Move test files around a bit
